+++ packages/be-ts-lib/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 🔖 [2.0.83] (2025-06-25)
+
+### ✅ Added
+
+- CHANGELOG.md file now included in published package
+
 ## 🔖 [2.0.82] (2025-06-19)
 
 ### 🏗️ Changed
@@ -20,4 +26,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 <!-- Link References -->
 
+[2.0.83]: https://github.com/aneuhold/ts-libs/compare/be-ts-lib-v2.0.82...be-ts-lib-v2.0.83
 [2.0.82]: https://github.com/aneuhold/ts-libs/releases/tag/be-ts-lib-v2.0.82